requirements.txt
asgiref==3.8.1
botocore==1.31.85
cement==2.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.3
Django==5.1
idna==3.8
jmespath==1.0.1
pathspec==0.10.1
psycopg2-binary==2.9.9
pypiwin32==223
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pywin32==306
PyYAML==6.0.2
requests==2.32.3
semantic-version==2.8.5
six==1.16.0
sqlparse==0.5.1
termcolor==1.1.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.19
wcwidth==0.1.9
whitenoise==6.7.0
